requirements.txt
flask
pandas
numpy
json
matplotlib
plotly-express
seaborn
scikit-learn